Fix #1907. Support single feature styling in OL3
