+++ web/client/components/map/openlayers/Feature.jsx
@@ -9,10 +9,13 @@
 var React = require('react');
 var ol = require('openlayers');
 const {isEqual} = require('lodash');
+const {getStyle} = require('./VectorStyle');
 
 let Feature = React.createClass({
     propTypes: {
         type: React.PropTypes.string,
+        layerStyle: React.PropTypes.object,
+        style: React.PropTypes.object,
         properties: React.PropTypes.object,
         crs: React.PropTypes.string,
         container: React.PropTypes.object, // TODO it must be a ol.layer.vector (maybe pass the source is more correct here?)
@@ -36,11 +39,19 @@ let Feature = React.createClass({
                 geometry: this.props.geometry,
                 id: this.props.msId});
             this._feature.forEach((f) => f.getGeometry().transform(this.props.featuresCrs, this.props.crs || 'EPSG:3857'));
+            if (this.props.style && (this.props.style !== this.props.layerStyle)) {
+                this._feature.forEach((f) => { f.setStyle(getStyle({style: this.props.style})); });
+            }
             this.props.container.getSource().addFeatures(this._feature);
         }
     },
-    componentWillReceiveProps(newProps) {
-        if (!isEqual(newProps.properties, this.props.properties) || !isEqual(newProps.geometry, this.props.geometry)) {
+    shouldComponentUpdate(nextProps) {
+        return !isEqual(nextProps.properties, this.props.properties)
+        || !isEqual(nextProps.geometry, this.props.geometry)
+        || !isEqual(nextProps.style, this.props.style);
+    },
+    componentWillUpdate(newProps) {
+        if (!isEqual(newProps.properties, this.props.properties) || !isEqual(newProps.geometry, this.props.geometry) || !isEqual(newProps.style, this.props.style)) {
             this.removeFromContainer();
             const format = new ol.format.GeoJSON();
             const geometry = newProps.geometry && newProps.geometry.coordinates;
@@ -48,12 +59,14 @@ let Feature = React.createClass({
             if (newProps.container && geometry) {
                 this._feature = format.readFeatures({type: newProps.type, properties: newProps.properties, geometry: newProps.geometry, id: this.props.msId});
                 this._feature.forEach((f) => f.getGeometry().transform(newProps.featuresCrs, this.props.crs || 'EPSG:3857'));
+                this._feature.forEach((f) => {
+                    if ((newProps.style !== this.props.style) && (newProps.layerStyle !== newProps.style)) {
+                        f.setStyle(getStyle({style: newProps.style}));
+                    }
+                });
                 newProps.container.getSource().addFeatures(this._feature);
             }
         }
-    },
-    shouldComponentUpdate(nextProps) {
-        return !isEqual(nextProps.properties, this.props.properties) || !isEqual(nextProps.geometry, this.props.geometry);
     },
     componentWillUnmount() {
         this.removeFromContainer();
